Modularising javascript, pay attention to AI and elements that are NOT transitioned. Complete these nextgit add .! Appy for jobs
