+++ css/otherStyles.css
@@ -1,3 +1,90 @@
+
+.sticky {
+    background-color: #ffffff;
+    opacity: 1;
+    top: 0px;
+  }
+  
+  a.nav-link {
+    color: #6e6f79;
+    transform: scale(1);
+    font-family: "Open Sans";
+    font-size: 16px;
+    transition: all 0.3s;
+    background-color: transparent;
+    opacity: 1;
+  }
+  
+  a.nav-link:hover {
+    color: #162a43;
+    transform: scale(1.15);
+    /* display: block;  */
+  }
+  
+  a.nav-link.active {
+    color: #162a43;
+    transform: scale(1.3);
+  }
+  
+  .dropdown-menu {
+    position: absolute;
+    top: 100%;
+    border-style: groove;
+    font-size: 14px;
+    background-color: transparent;
+    text-align: left;
+    inline-size: auto;
+    border-width: auto;
+    border-radius: 5px;
+    border-right: none;
+    border-left: none;
+  }
+  
+  a.dropdown-item {
+    color: #fdfdff;
+    font-size: 0.9em;
+    padding: 2px 0px 1px 0px;
+  }
+  
+  a.dropdown-item:hover {
+    color: #162a43;
+    transform: scale(1.1);
+    background-color: transparent !important;
+  }
+  
+  a.dropdown-item:active {
+    transform: scale(1.2);
+    color: #162a43;
+    background-color: transparent !important;
+  }
+  
+  .dropdown .dropdown-menu .navbar-nav {
+    --bs-nav-link-hover-colour: #162a43;
+  }
+  
+  ul.navbar-nav {
+    position: absolute;
+    align-self: center;
+  }
+  
+  .navbar {
+    --bs-navbar-active-color: #162a43;
+    --bs-navbar-brand-color: #162a43;
+    --bs-navbar-brand-hover-color: #162a43;
+    position: fixed;
+    top: 0px;
+    max-height: 50px;
+    width: 100%;
+    padding-top: 30px;
+    background: none;
+    overflow: visible;
+    -webkit-transition: all 1s;
+    transition: all 1s;
+    opacity: 0;
+    top: -700px;
+    padding-bottom: 20px;
+  }
+
 .slider, 
 .slider > div {
     /* Images default to Center Center. Maybe try 'center top'? */
