+++ index.html
@@ -13,7 +13,8 @@
   <!-- <link rel="stylesheet" href="css/test.css"> -->
 
   <script src="https://code.jquery.com/jquery-1.11.0.min.js" defer></script>
-  <script defer src="js/main.js" defer></script>
+  
+  <script type="module" src="js/main.js" defer></script>
   <!-- <script src="js/test.js" defer></script> -->
   <!-- <script src="js/testTwo.js" defer></script> -->
 
@@ -112,21 +113,21 @@
         <image class="img-fluid" id="cloud5" xlink:href="assets/cl4.webp" width="1200" height="800" />
         <image class="img-fluid" id="cloud1" xlink:href="assets/cl1.webp" width="1200" height="800" />
 
-        <text class="lrgtext" id="see" fill="#fff" dominant-baseline="middle" text-anchor="middle"
+        <text class="lrgtext" id="see" fill="#fff"  dominant-baseline="middle" text-anchor="middle"
           style="visibility: hidden;" onclick="autoScrollNow()">SEE</text>
 
         <g mask="url(#m)">
           <rect fill="#fff" width="100%" height="100%" />
           <rect id="meshaker" />
           <a class="meLink" onclick="showStatementContact();">
-            <text class="lrgtext" id="me" fill="#162a43" dominant-baseline="middle" text-anchor="middle"
+            <text class="lrgtext" id="me" fill="#162a43" aria-disabled="true" dominant-baseline="middle" text-anchor="middle"
               onclick="autoScrollNow()">ME</text>
           </a>
           <div class="scroll-arrow" id="down" x="455" y="330" onclick="autoScrollNow()"></div>
         </g>
     </div>
 
-    <div class="container-fluid thumbnails" id="thumbnails" style="visibility: hidden;">
+    <div class="container-fluid thumbnails" id="thumbnails">
       <!-- href="software.html" -->
       <a class="thumbs">
         <div class="thumbShape thumbs" id="software" onclick="showWip();">
