+++ photography.html
@@ -1,8 +1,4 @@
-<!-- Building on the shoulders of giants, special thanks to:
-https://codepen.io/EricPorter/pen/XaOEpP
 
-
- -->
 
 <!DOCTYPE html>
   <html lang="en">
@@ -113,10 +109,13 @@ https://codepen.io/EricPorter/pen/XaOEpP
           </div>
         </div>
       </nav>
+      
   
   <!-- Slider 1 -->
 <div class="slider" id="slider1">
   <!-- Slides -->
+  
+  <h2>ONE</h2>
   <div style="background-image:url(https://s3-us-west-2.amazonaws.com/s.cdpn.io/30256/1200_110627-8240-Myst.jpg)"></div>
   <div style="background-image:url(https://s3-us-west-2.amazonaws.com/s.cdpn.io/30256/Gif-pont-lueur-600.gif)"></div>
   <div style="background-image:url(https://s3-us-west-2.amazonaws.com/s.cdpn.io/30256/1200_Sharpened-version.jpg)"></div>
@@ -140,6 +139,7 @@ https://codepen.io/EricPorter/pen/XaOEpP
 <div class="row2Wrap">
   <!-- Slider 2 -->
   <div class="slider" id="slider2">
+      <h2>TWO</h2>
 
       <div style="background-image:url(https://s3-us-west-2.amazonaws.com/s.cdpn.io/30256/blurredBeachGrassCloseup.jpg)"></div>
       <div style="background-image:url(https://s3-us-west-2.amazonaws.com/s.cdpn.io/30256/1200_110627-8240-Myst.jpg)"></div>
@@ -156,32 +156,33 @@ https://codepen.io/EricPorter/pen/XaOEpP
               <path d="M 10,50 L 60,100 L 70,90 L 30,50  L 70,10 L 60,0 Z" transform="translate(100, 100) rotate(180) "></path>
           </svg>
       </i>
-      <!-- Title Bar -->
-      <!--     
+      Title Bar
+          
       <span class="titleBar"> 
           <h1>I am like a leaf in the wind.</h1> 
           <p>Watch me soar!</p>
       </span> 
-      -->
+     
   </div>
 
   <!-- <h3>Individual slide text</h3> -->
 
   <!-- Slider 3 -->
   <div class="slider" id="slider3">
+      
       <div style="background-image:url(https://s3-us-west-2.amazonaws.com/s.cdpn.io/30256/jungle.jpg)">
           <span>
-              <h2>I'm a title for Slide #1</h2>
+              <h2>ONE</h2>
           </span>
       </div>
       <div style="background-image:url(https://s3-us-west-2.amazonaws.com/s.cdpn.io/30256/1200_110627-8240-Myst.jpg)">
           <span>
-              <h2>I'm a title for Slide #2</h2>
+              <h2>TWO</h2>
           </span>
       </div>
       <div style="background-image:url(https://s3-us-west-2.amazonaws.com/s.cdpn.io/30256/1200_bodie-11.jpg)">
           <span>
-              <h2>I'm a title for Slide #3</h2>
+              <h2>THREE</h2>
           </span>
       </div>
       <!-- The Arrows -->
@@ -192,11 +193,11 @@ https://codepen.io/EricPorter/pen/XaOEpP
           <svg viewBox="0 0 100 100">
               <path d="M 10,50 L 60,100 L 70,90 L 30,50  L 70,10 L 60,0 Z" transform="translate(100, 100) rotate(180) "></path>
           </svg></i>
-      <!-- Title Bar -->
-      <!--     <span class="titleBar">
+      Title Bar
+          <span class="titleBar">
       <h1>I am like a leaf in the wind.</h1> 
       <p>Watch me soar!</p>
-  </span> -->
+  </span>
   </div>
 
 </div>
@@ -216,7 +217,6 @@ https://codepen.io/EricPorter/pen/XaOEpP
       </ul>
   </h3>
 </div>
-
